Use Protobuf for parse tree format, instead of manually generated JSON
Up until now, this library relied on JSON to pass the parse result back
to the caller, which has a number of downsides, most importantly that
we don't have a readily available parser for JSON thats not tied to a
running Postgres server. That in turn makes it hard to provide
cross-language features such as deparsing directly in this library
(which would require reading back a parse tree that gets passed in).

Protobuf isn't perfect, but its straightforward enough to generate the
schema definitions for the parse tree nodes, and output the tree using
the C++ protobuf library. Note that this also means the library now
depends on the C++ standard library (not just the C standard library),
but this is thought to not be a problem in practical uses of this library.

Another benefit of the Protobuf C++ library is that it can output JSON
as well, which looks similar to the previously generated JSON. For callers
that prefer a JSON output, this (almost identical) JSON format is also
available. Note that deparser use cases will likely require the use of
the binary Protobuf format, which is the recommended format for downstream
libraries to adopt.