Separate split function into two: scanner and parser
These behave differently and can both be valuable, most importantly
when splitting statements in a file that may contain syntax errors that
cause a parser error, but are accepted by the scanner.

This particular situation is the case with the Postgres regression tests,
which is why we need this functionality for the deparser tests.