Correctly set non-standard Enum values when reading back protobuf
We previously assumed that all enums are zero-indexed and incrementing,
but that is not the case for some of the enum values, and this leads
to silent errors when reading back the parse trees.

Note that protobuf still ensures all enums are zero-indexed and
incrementing after that, because the protobuf format is best used that
way.